+++ CotEditor/Sources/Text Finder/Multiple Replace/Views/MultipleReplaceSplitView.swift
@@ -49,7 +49,7 @@ struct MultipleReplaceSplitView: View {
                 self.saveSetting()
             }
         }
-        .onChange(of: self.selection) { _, newValue in
+        .onChange(of: self.selection, initial: true) { _, newValue in
             guard let newValue else { return }
             self.changeSetting(to: newValue)
         }